refactor: Move StatisticEntity to report module and simplify
Remove StatisticId and reportId dependencies from StatisticEntity
Streamline entity creation by removing unnecessary ID generation
