+++ src/domain/entities/report/statistic.entity.ts
@@ -1,45 +1,22 @@
-import { ReportId } from '@/domain/entities/report/report-id.js';
-import { StatisticId } from '@/domain/entities/statistic/statistic-id.js';
-
 interface StatisticEntityProps {
-  id: StatisticId;
-  reportId: ReportId;
   promptTokens: number;
   completionTokens: number;
-  createdAt: Date;
 }
 
 export class StatisticEntity {
-  private readonly _id: StatisticId;
-
-  private readonly _reportId: ReportId;
-
   private _promptTokens: number;
 
   private _completionTokens: number;
 
-  private readonly _createdAt: Date;
-
   private constructor(props: StatisticEntityProps) {
-    this._id = props.id;
-    this._reportId = props.reportId;
     this._promptTokens = props.promptTokens;
     this._completionTokens = props.completionTokens;
-    this._createdAt = props.createdAt;
   }
 
-  static create(payload: {
-    id: StatisticId;
-    reportId: ReportId
-  }): StatisticEntity {
-    const { id, reportId } = payload;
-
+  public static create(): StatisticEntity {
     return new StatisticEntity({
-      id,
-      reportId,
       promptTokens: 0,
       completionTokens: 0,
-      createdAt: new Date(),
     });
   }
